+++ subsys/bluetooth/controller/ll_sw/nrf.cmake
@@ -80,7 +80,7 @@ if(CONFIG_BT_LL_SW_SPLIT)
     CONFIG_BT_CTLR_PERIPHERAL_ISO
     ll_sw/nordic/lll/lll_peripheral_iso.c
     )
-  if(CONFIG_BT_CTLR_ISO)
+  if(CONFIG_BT_CTLR_ISO_VENDOR_DATA_PATH)
     zephyr_library_sources(
       ll_sw/nordic/ull/ull_iso_vendor.c
       )
